+++ app/src/main/java/de/baumann/hhsmoodle/fragmentsMain/FragmentNotes.java
@@ -47,6 +47,8 @@ import android.widget.ListView;
 import android.widget.SimpleAdapter;
 import android.widget.TextView;
 
+import net.sqlcipher.database.SQLiteDatabase;
+
 import java.io.File;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -263,6 +265,7 @@ public class FragmentNotes extends Fragment {
             }
         });
 
+        SQLiteDatabase.loadLibs(getActivity());
         setNotesList();
         return rootView;
     }