+++ app/src/main/java/de/baumann/hhsmoodle/helper/Database_Browser.java
@@ -22,10 +22,10 @@ package de.baumann.hhsmoodle.helper;
 import android.content.Context;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager.NameNotFoundException;
-import android.database.Cursor;
-import android.database.sqlite.SQLiteDatabase;
-import android.database.sqlite.SQLiteOpenHelper;
-import android.database.sqlite.SQLiteStatement;
+import net.sqlcipher.Cursor;
+import net.sqlcipher.database.SQLiteDatabase;
+import net.sqlcipher.database.SQLiteOpenHelper;
+import net.sqlcipher.database.SQLiteStatement;
 import android.preference.PreferenceManager;
 
 import java.util.ArrayList;
@@ -36,7 +36,7 @@ public class Database_Browser extends SQLiteOpenHelper {
     public Database_Browser(Context context)
             throws NameNotFoundException {
         super(context,
-                "browser.db",
+                "browser_encrypted.db",
                 null,
                 context.getPackageManager().getPackageInfo(context.getPackageName(), 0).versionCode);
     }
@@ -64,7 +64,7 @@ public class Database_Browser extends SQLiteOpenHelper {
     public void loadInitialData() {
         int seqno = 0;
 
-        SQLiteDatabase db = getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase("oOk_-58F,iNEdJ)Fv3}H}]kMh]");
         db.beginTransaction();
 
         SQLiteStatement stmt = db.compileStatement("INSERT INTO bookmarks VALUES(?, ?, ?, ?)");
@@ -80,7 +80,7 @@ public class Database_Browser extends SQLiteOpenHelper {
     }
 
     public int getRecordCount() {
-        SQLiteDatabase db = getReadableDatabase();
+        SQLiteDatabase db = getReadableDatabase("oOk_-58F,iNEdJ)Fv3}H}]kMh]");
 
         int ret = 0;
 
@@ -97,7 +97,7 @@ public class Database_Browser extends SQLiteOpenHelper {
     }
 
     public void getBookmarks(ArrayList<String[]> data, Context context) {
-        SQLiteDatabase db = getReadableDatabase();
+        SQLiteDatabase db = getReadableDatabase("oOk_-58F,iNEdJ)Fv3}H}]kMh]");
 
         PreferenceManager.setDefaultValues(context, R.xml.user_settings, false);
         SharedPreferences sp = PreferenceManager.getDefaultSharedPreferences(context);
@@ -152,7 +152,7 @@ public class Database_Browser extends SQLiteOpenHelper {
     public void addBookmark(String title, String url, String icon) {
         int seqno;
 
-        SQLiteDatabase db = getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase("oOk_-58F,iNEdJ)Fv3}H}]kMh]");
 
         String sql = "SELECT MAX(seqno) FROM bookmarks";
         Cursor c = db.rawQuery(sql, null);
@@ -177,7 +177,7 @@ public class Database_Browser extends SQLiteOpenHelper {
 
     public void deleteBookmark(int seqno) {
 
-        SQLiteDatabase db = getWritableDatabase();
+        SQLiteDatabase db = getWritableDatabase("oOk_-58F,iNEdJ)Fv3}H}]kMh]");
         db.beginTransaction();
 
         SQLiteStatement stmt = db.compileStatement("DELETE FROM bookmarks WHERE seqno = ?");